requirements.txt
pandas~=2.2.3
sqlalchemy~=2.0.23
yaml~=0.2.5
pyyaml~=6.0.1
tqdm~=4.67.1
google-api-python-client~=2.154.0

polars~=1.15.0
matplotlib~=3.9.2
numpy~=2.1.3
plotly~=5.24.1
scikit-learn~=1.5.2
protobuf~=4.25.3
google-auth-oauthlib~=1.2.1